Differentiating ToggleButton and ToggleInput to their own files
