+++ src/components/Form/Toggle.tsx
@@ -5,14 +5,13 @@ import React, {
   ComponentClass,
   FunctionComponent,
 } from 'react';
-import { HtmlLabel, HtmlInput, HtmlSpan } from '../../reset';
 import { idGenerator } from '../../utils/uuid';
-import classnames from 'classnames';
+import { ToggleInput } from './ToggleInput';
+import { ToggleButton } from './ToggleButton';
+export { ToggleButton, ToggleInput };
 
-const baseClassName = 'fi-toggle';
-const toggleDisabledClassName = `${baseClassName}--disabled`;
-const toggleInputClassName = `${baseClassName}_input`;
-const toggleLabelClassName = `${baseClassName}_label`;
+export const baseClassName = 'fi-toggle';
+// const toggleDisabledClassName = `${baseClassName}--disabled`;
 
 export interface ToggleInputProps {
   /** State of input checkbox */
@@ -24,20 +23,6 @@ export interface ToggleInputProps {
   controlled?: boolean;
   'aria-label'?: string;
   'aria-labelledby'?: string;
-}
-
-export class ToggleInput extends Component<ToggleInputProps> {
-  render() {
-    const { disabled = false, controlled, checked, ...passProps } = this.props;
-    return (
-      <HtmlInput
-        disabled={disabled}
-        {...passProps}
-        type="checkbox"
-        checked={!!checked}
-      />
-    );
-  }
 }
 
 export interface ToggleProps {
@@ -57,6 +42,8 @@ export interface ToggleProps {
    * Label element content
    */
   children?: ReactNode;
+  /** Use input instead of button */
+  withInput?: boolean;
   /** Pass custom props to Toggle's input component/element */
   toggleInputProps?: ToggleInputProps;
   /** Customized ToggleInput-component */
@@ -76,97 +63,52 @@ export interface ToggleProps {
   id?: string;
 }
 
-interface ToggleState {
+export interface ToggleState {
   toggleState: boolean;
 }
 
-const componentOrElementWithProps = (
-  component: ReactElement<any> | FunctionComponent<any> | ComponentClass<any>,
-  props: object,
-) => {
-  if (!!component) {
-    if (React.isValidElement(component)) {
-      return React.cloneElement(component, props); // element
-    }
-    return React.createElement(
-      component as FunctionComponent<any> | ComponentClass<any>,
-      props,
-    ); // component
-  }
-  return;
-};
-
 export class Toggle extends Component<ToggleProps> {
-  state: ToggleState = {
-    toggleState: !!this.props.checked || !!this.props.defaultChecked || false,
-  };
+  // state: ToggleState = {
+  //   toggleState: !!this.props.checked || !!this.props.defaultChecked || false,
+  // };
 
-  componentWillReceiveProps(nextProps: ToggleProps) {
-    const { checked } = nextProps;
-    if (!!checked) {
-      this.setState({ toggleState: !!checked });
-    }
-  }
+  // componentWillReceiveProps(nextProps: ToggleProps) {
+  //   const { checked } = nextProps;
+  //   if (!!checked) {
+  //     this.setState({ toggleState: !!checked });
+  //   }
+  // }
 
-  handleClick = () => {
-    const { checked, onClick } = this.props;
-    const { toggleState } = this.state;
-    if (checked === undefined) {
-      this.setState({ toggleState: !toggleState });
-    }
-    if (!!onClick) {
-      onClick({ toggleState: !toggleState });
-    }
-  };
+  // handleClick = () => {
+  //   const { checked, onClick } = this.props;
+  //   const { toggleState } = this.state;
+  //   if (checked === undefined) {
+  //     this.setState({ toggleState: !toggleState });
+  //   }
+  //   if (!!onClick) {
+  //     onClick({ toggleState: !toggleState });
+  //   }
+  // };
 
   render() {
-    const {
-      disabled,
-      className,
-      children,
-      toggleInputProps,
-      toggleInputComponent,
-      'aria-label': ariaLabel,
-      'aria-labelledby': ariaLabelledBy,
-      checked: dissMissChecked,
-      defaultChecked: dissMissDefaultChecked,
-      onClick: dissMissOnClick,
-      id: propId,
-      ...passProps
-    } = this.props;
-    const { toggleState } = this.state;
+    const { withInput, id: propId, ...passProps } = this.props;
+
     const id = idGenerator(propId);
-    const newToggleInputProps = {
-      disabled,
-      'aria-label': ariaLabel,
-      'aria-labelledby': ariaLabelledBy,
-      checked: !!toggleState,
-      className: toggleInputClassName,
-      onChange: this.handleClick,
-      ...toggleInputProps,
+
+    // TODO: passaa kaikki propsit, lisää mahdollisesti jotain; esim. id ja ..?
+    const newToggleProps = {
       id,
+      ...passProps,
     };
 
     return (
-      <HtmlSpan
-        className={classnames(className, baseClassName, {
-          [toggleDisabledClassName]: !!disabled,
-        })}
-      >
-        {!!toggleInputComponent ? (
-          componentOrElementWithProps(toggleInputComponent, newToggleInputProps) // element
+      <React.Fragment>
+        {!!withInput ? (
+          <ToggleInput {...newToggleProps} />
         ) : (
-          <ToggleInput {...newToggleInputProps} />
+          <ToggleButton {...newToggleProps} />
         )}
-        <HtmlLabel
-          {...passProps}
-          className={toggleLabelClassName}
-          onClick={this.handleClick}
-          htmlFor={id}
-        >
-          {children}
-        </HtmlLabel>
-      </HtmlSpan>
+      </React.Fragment>
     );
   }
 }
